feat: 댓글 등록/수정/삭제 인증 로직 추가

+++ src/main/java/NotFound/next_campus/domain/comment/service/CommentServiceImpl.java
@@ -10,6 +10,7 @@ import NotFound.next_campus.domain.member.model.Role;
 import NotFound.next_campus.domain.member.repository.MemberRepository;
 import NotFound.next_campus.domain.post.model.Post;
 import NotFound.next_campus.domain.post.repository.PostRepository;
+import NotFound.next_campus.global.auth.user.CustomUserDetails;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -27,10 +28,9 @@ public class CommentServiceImpl implements CommentService {
     private final CommentRepository commentRepository;
 
     @Override
-    public Long saveComment(CreateCommentDTO dto) {
+    public Long saveComment(CreateCommentDTO dto, CustomUserDetails userDetails) {
 
-        Member member = memberRepository.findById(dto.getMemberId())
-                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 사용자입니다."));
+        Member member = userDetails.getMember();
         Post post = postRepository.findById(dto.getPostId())
                 .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 게시물입니다."));
 
@@ -44,15 +44,15 @@ public class CommentServiceImpl implements CommentService {
     }
 
     @Override
-    public void updateComment(Long commentId, UpdateCommentDTO dto) {
+    public void updateComment(Long commentId, UpdateCommentDTO dto, CustomUserDetails userDetails) {
 
-        Member member = memberRepository.findById(dto.getMemberId())
-                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 사용자입니다."));
+        Member member = userDetails.getMember();
         Comment comment = commentRepository.findById(commentId)
                 .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 댓글입니다."));
 
-        if (Role.USER.equals(member.getRole())
-                && !comment.getMember().equals(member)) {
+        // 본인이 등록한 댓글이거나 관리자인 경우에만 수정 가능
+        if (!comment.getMember().equals(member) &&
+                !Role.ADMIN.equals(member.getRole())) {
             throw new IllegalArgumentException("해당 댓글에 대한 수정 권한이 없습니다.");
         }
 
@@ -60,12 +60,16 @@ public class CommentServiceImpl implements CommentService {
     }
 
     @Override
-    public void deleteComment(Long commentId) {
-
-        /* 권한 체크 로직 추가 */
+    public void deleteComment(Long commentId, CustomUserDetails userDetails) {
 
         Comment comment = commentRepository.findById(commentId)
                 .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 댓글입니다."));
+
+        // 본인이 등록한 댓글이거나 관리자인 경우에만 삭제 가능
+        if (!comment.getMember().equals(userDetails.getMember()) &&
+                !Role.ADMIN.equals(userDetails.getRole())) {
+            throw new IllegalArgumentException("해당 댓글에 대한 삭제 권한이 없습니다.");
+        }
 
         commentRepository.delete(comment);
     }
